fix(mz-cloud): 变动常量文件，并移除错误的路径正则

+++ mz-modules/mz-system/mz-system-model/src/main/java/com/mz/system/model/entity/SysConfigEntity.java
@@ -32,7 +32,7 @@ public class SysConfigEntity extends BaseEntity implements Serializable {
 	 */
     @ApiModelProperty("参数主键")
 	@TableId(type = IdType.AUTO)
-	private Integer configId;
+	private Long configId;
 	/**
 	 * 参数名称
 	 */
